requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.10
gunicorn==20.1.0
Flask-Login==0.6.3
scikit-learn==1.5.1
pandas==2.2.2
gunicorn==20.1.0